Moved around Io enums
