+++ src/Bin/OpenEnroth/OpenEnroth.cpp
@@ -18,6 +18,7 @@
 #include "Engine/Engine.h"
 
 #include "Io/KeyboardActionMapping.h"
+#include "Io/InputEnumFunctions.h"
 
 #include "Library/StackTrace/StackTraceOnCrash.h"
 #include "Library/Platform/Application/PlatformApplication.h"
@@ -72,7 +73,7 @@ void migrateTrace(OpenEnrothOptions::Migration migration, EventTrace *trace) {
     case OpenEnrothOptions::MIGRATION_COLLAPSE_KEY_EVENTS:
         std::unordered_set<PlatformKey> keys;
         for (InputAction inputAction : allInputActions())
-            if (keyboardActionMapping->toggleTypeFor(inputAction) != TOGGLE_ONCE)
+            if (toggleTypeForInputAction(inputAction) != TOGGLE_ONCE)
                 keys.insert(keyboardActionMapping->keyFor(inputAction));
         return EventTrace::migrateCollapseKeyEvents(keys, trace);
     }
