+++ src/GUI/GUIWindow.cpp
@@ -39,7 +39,7 @@
 #include "GUI/UI/UIPopup.h"
 #include "GUI/UI/UIDialogue.h"
 
-#include "Io/InputAction.h"
+#include "Io/InputEnums.h"
 #include "Io/KeyboardInputHandler.h"
 #include "Io/Mouse.h"
