+++ src/Io/InputEnumFunctions.cpp
@@ -1,0 +1,318 @@
+#include "InputEnumFunctions.h"
+
+#include <string>
+
+#include "Library/Serialization/EnumSerialization.h"
+
+#include "Utility/IndexedArray.h"
+
+// TODO(captainurist): these should be localizable.
+MM_DEFINE_ENUM_SERIALIZATION_FUNCTIONS(InputAction, CASE_INSENSITIVE, {
+    {INPUT_ACTION_MOVE_FORWARD, "FORWARD"},
+    {INPUT_ACTION_MOVE_BACKWARDS, "BACKWARD"},
+    {INPUT_ACTION_TURN_LEFT, "LEFT"},
+    {INPUT_ACTION_TURN_RIGHT, "RIGHT"},
+    {INPUT_ACTION_YELL, "YELL"},
+    {INPUT_ACTION_JUMP, "JUMP"},
+    {INPUT_ACTION_TOGGLE_TURN_BASED, "COMBAT"},
+    {INPUT_ACTION_QUICK_CAST, "CAST READY"},
+    {INPUT_ACTION_ATTACK, "ATTACK"},
+    {INPUT_ACTION_INTERACT, "TRIGGER"},
+    {INPUT_ACTION_OPEN_SPELLBOOK, "CAST"},
+    {INPUT_ACTION_PASS, "PASS"},
+    {INPUT_ACTION_NEXT_CHAR, "CHAR CYCLE"},
+    {INPUT_ACTION_OPEN_QUESTS, "QUEST"},
+    {INPUT_ACTION_OPEN_QUICK_REFERENCE, "QUICK REF"},
+    {INPUT_ACTION_REST, "REST"},
+    {INPUT_ACTION_OPEN_CALENDAR, "TIME/CAL"},
+    {INPUT_ACTION_OPEN_AUTONOTES, "AUTONOTES"},
+    {INPUT_ACTION_OPEN_MAP, "MAP BOOK"},
+    {INPUT_ACTION_TOGGLE_ALWAYS_RUN, "ALWAYS RUN"},
+    {INPUT_ACTION_LOOK_UP, "LOOK UP"},
+    {INPUT_ACTION_LOOK_DOWN, "LOOK DOWN"},
+    {INPUT_ACTION_CENTER_VIEW, "CTR VIEW"},
+    {INPUT_ACTION_ZOOM_IN, "ZOOM IN"},
+    {INPUT_ACTION_ZOOM_OUT, "ZOOM OUT"},
+    {INPUT_ACTION_FLY_UP, "FLY UP"},
+    {INPUT_ACTION_FLY_DOWN, "FLY DOWN"},
+    {INPUT_ACTION_FLY_LAND, "LAND"},
+    {INPUT_ACTION_STRAFE_LEFT, "STRAFE LEFT"},
+    {INPUT_ACTION_STRAFE_RIGHT, "STRAFE RIGHT"},
+    {INPUT_ACTION_QUICK_SAVE, "QUICK SAVE"},
+    {INPUT_ACTION_QUICK_LOAD, "QUICK LOAD"},
+    {INPUT_ACTION_OPEN_HISTORY, "HISTORY"},
+    {INPUT_ACTION_OPEN_STATS, "STATS"},
+    {INPUT_ACTION_OPEN_SKILLS, "SKILLS"},
+    {INPUT_ACTION_OPEN_AWARDS, "INVENTORY"},
+    {INPUT_ACTION_OPEN_INVENTORY, "AWARDS"},
+    {INPUT_ACTION_NEW_GAME, "NEW GAME"},
+    {INPUT_ACTION_SAVE_GAME, "SAVE GAME"},
+    {INPUT_ACTION_LOAD_GAME, "Load GAME"},
+    {INPUT_ACTION_EXIT_GAME, "EXIT GAME"},
+    {INPUT_ACTION_BACK_TO_GAME, "RETURN TO GAME"},
+    {INPUT_ACTION_OPEN_CONTROLS, "CONTROLS"},
+    {INPUT_ACTION_OPEN_OPTIONS, "OPTIONS"},
+    {INPUT_ACTION_SHOW_CREDITS, "CREDITS"},
+    {INPUT_ACTION_PARTY_CREATION_CLEAR, "CLEAR"},
+    {INPUT_ACTION_PARTY_CREATION_DONE, "RETURN"},
+    {INPUT_ACTION_PARTY_CREATION_DEC, "MINUS"},
+    {INPUT_ACTION_PARTY_CREATION_INC, "PLUS"},
+    {INPUT_ACTION_TRANSITION_YES, "YES"},
+    {INPUT_ACTION_TRANSITION_NO, "NO"},
+    {INPUT_ACTION_REST_HEAL, "REST 8 HOURS"},
+    {INPUT_ACTION_REST_WAIT_TILL_DAWN, "WAIT TILL DAWN"},
+    {INPUT_ACTION_REST_WAIT_1_HOUR, "WAIT 1 HOUR"},
+    {INPUT_ACTION_REST_WAIT_5_MINUTES, "WAIT 5 MINUTES"},
+    {INPUT_ACTION_TAKE_SCREENSHOT, "SCREENSHOT"},
+    {INPUT_ACTION_OPEN_CONSOLE, "CONSOLE"},
+    {INPUT_ACTION_SELECT_CHAR_1, "CHAR 1"},
+    {INPUT_ACTION_SELECT_CHAR_2, "CHAR 2"},
+    {INPUT_ACTION_SELECT_CHAR_3, "CHAR 3"},
+    {INPUT_ACTION_SELECT_CHAR_4, "CHAR 4"},
+    {INPUT_ACTION_SELECT_NPC_1, "NPC 1"},
+    {INPUT_ACTION_SELECT_NPC_2, "NPC 2"},
+    {INPUT_ACTION_DIALOG_UP, "DIALOG UP"},
+    {INPUT_ACTION_DIALOG_DOWN, "DIALOG DOWN"},
+    {INPUT_ACTION_DIALOG_LEFT, "DIALOG LEFT"},
+    {INPUT_ACTION_DIALOG_RIGHT, "DIALOG RIGHT"},
+    {INPUT_ACTION_DIALOG_PRESS, "DIALOG SELECT"},
+    {INPUT_ACTION_ESCAPE, "ESCAPE"},
+    {INPUT_ACTION_TOGGLE_MOUSE_LOOK, "TOGGLE MOUSE LOOK"},
+    {INPUT_ACTION_TOGGLE_WINDOW_MODE, "TOGGLE WINDOW MODE"}
+});
+
+std::string GetDisplayName(InputAction action) {
+    std::string result;
+    if (trySerialize(action, &result))
+        return result;
+    return "-INVALID-ACTION-";
+}
+
+static constexpr IndexedArray<KeyToggleType, INPUT_ACTION_FIRST_VALID, INPUT_ACTION_LAST_VALID> toggleTypeByInputAction = {
+    {INPUT_ACTION_MOVE_FORWARD,         TOGGLE_CONTINUOUSLY},
+    {INPUT_ACTION_MOVE_BACKWARDS,       TOGGLE_CONTINUOUSLY},
+    {INPUT_ACTION_TURN_LEFT,            TOGGLE_CONTINUOUSLY},
+    {INPUT_ACTION_TURN_RIGHT,           TOGGLE_CONTINUOUSLY},
+    {INPUT_ACTION_ATTACK,               TOGGLE_CONTINUOUSLY_WITH_DELAY},
+    {INPUT_ACTION_QUICK_CAST,           TOGGLE_CONTINUOUSLY_WITH_DELAY},
+    {INPUT_ACTION_YELL,                 TOGGLE_CONTINUOUSLY_WITH_DELAY},
+    {INPUT_ACTION_JUMP,                 TOGGLE_ONCE}, // TODO: TOGGLE_Continuously
+    {INPUT_ACTION_TOGGLE_TURN_BASED,    TOGGLE_ONCE},
+    {INPUT_ACTION_INTERACT,             TOGGLE_ONCE}, // TODO: TOGGLE_DelayContinuous
+    {INPUT_ACTION_OPEN_SPELLBOOK,       TOGGLE_ONCE},
+    {INPUT_ACTION_PASS,                 TOGGLE_CONTINUOUSLY_WITH_DELAY},
+    {INPUT_ACTION_NEXT_CHAR,            TOGGLE_CONTINUOUSLY_WITH_DELAY},
+    {INPUT_ACTION_OPEN_QUESTS,          TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_QUICK_REFERENCE, TOGGLE_ONCE},
+    {INPUT_ACTION_REST,                 TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_CALENDAR,        TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_AUTONOTES,       TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_MAP,             TOGGLE_ONCE},
+    {INPUT_ACTION_LOOK_UP,              TOGGLE_ONCE}, // TODO: TOGGLE_DelayContinuous
+    {INPUT_ACTION_LOOK_DOWN,            TOGGLE_ONCE}, // TODO: TOGGLE_DelayContinuous
+    {INPUT_ACTION_CENTER_VIEW,          TOGGLE_ONCE},
+    {INPUT_ACTION_ZOOM_IN,              TOGGLE_CONTINUOUSLY_WITH_DELAY},
+    {INPUT_ACTION_ZOOM_OUT,             TOGGLE_CONTINUOUSLY_WITH_DELAY},
+    {INPUT_ACTION_FLY_UP,               TOGGLE_CONTINUOUSLY},
+    {INPUT_ACTION_FLY_DOWN,             TOGGLE_CONTINUOUSLY},
+    {INPUT_ACTION_FLY_LAND,             TOGGLE_ONCE},
+    {INPUT_ACTION_TOGGLE_ALWAYS_RUN,    TOGGLE_ONCE},
+    {INPUT_ACTION_STRAFE_LEFT,          TOGGLE_CONTINUOUSLY},
+    {INPUT_ACTION_STRAFE_RIGHT,         TOGGLE_CONTINUOUSLY},
+    {INPUT_ACTION_QUICK_SAVE,           TOGGLE_ONCE},
+    {INPUT_ACTION_QUICK_LOAD,           TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_HISTORY,         TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_STATS,           TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_SKILLS,          TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_INVENTORY,       TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_AWARDS,          TOGGLE_ONCE},
+    {INPUT_ACTION_NEW_GAME,             TOGGLE_ONCE},
+    {INPUT_ACTION_SAVE_GAME,            TOGGLE_ONCE},
+    {INPUT_ACTION_LOAD_GAME,            TOGGLE_ONCE},
+    {INPUT_ACTION_EXIT_GAME,            TOGGLE_ONCE},
+    {INPUT_ACTION_BACK_TO_GAME,         TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_CONTROLS,        TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_OPTIONS,         TOGGLE_ONCE},
+    {INPUT_ACTION_SHOW_CREDITS,         TOGGLE_ONCE},
+    {INPUT_ACTION_PARTY_CREATION_CLEAR, TOGGLE_ONCE},
+    {INPUT_ACTION_PARTY_CREATION_DONE,  TOGGLE_ONCE},
+    {INPUT_ACTION_PARTY_CREATION_DEC,   TOGGLE_ONCE},
+    {INPUT_ACTION_PARTY_CREATION_INC,   TOGGLE_ONCE},
+    {INPUT_ACTION_TRANSITION_YES,       TOGGLE_ONCE},
+    {INPUT_ACTION_TRANSITION_NO,        TOGGLE_ONCE},
+    {INPUT_ACTION_REST_HEAL,            TOGGLE_ONCE},
+    {INPUT_ACTION_REST_WAIT_TILL_DAWN,  TOGGLE_ONCE},
+    {INPUT_ACTION_REST_WAIT_1_HOUR,     TOGGLE_ONCE},
+    {INPUT_ACTION_REST_WAIT_5_MINUTES,  TOGGLE_ONCE},
+    {INPUT_ACTION_TAKE_SCREENSHOT,      TOGGLE_ONCE},
+    {INPUT_ACTION_OPEN_CONSOLE,         TOGGLE_ONCE},
+    {INPUT_ACTION_SELECT_CHAR_1,        TOGGLE_ONCE},
+    {INPUT_ACTION_SELECT_CHAR_2,        TOGGLE_ONCE},
+    {INPUT_ACTION_SELECT_CHAR_3,        TOGGLE_ONCE},
+    {INPUT_ACTION_SELECT_CHAR_4,        TOGGLE_ONCE},
+    {INPUT_ACTION_SELECT_NPC_1,         TOGGLE_ONCE},
+    {INPUT_ACTION_SELECT_NPC_2,         TOGGLE_ONCE},
+    {INPUT_ACTION_DIALOG_UP,            TOGGLE_ONCE},
+    {INPUT_ACTION_DIALOG_DOWN,          TOGGLE_ONCE},
+    {INPUT_ACTION_DIALOG_LEFT,          TOGGLE_ONCE},
+    {INPUT_ACTION_DIALOG_RIGHT,         TOGGLE_ONCE},
+    {INPUT_ACTION_DIALOG_PRESS,         TOGGLE_ONCE},
+    {INPUT_ACTION_ESCAPE,               TOGGLE_ONCE},
+    {INPUT_ACTION_TOGGLE_MOUSE_LOOK,    TOGGLE_ONCE},
+    {INPUT_ACTION_TOGGLE_WINDOW_MODE,   TOGGLE_CONTINUOUSLY},
+};
+
+KeyToggleType toggleTypeForInputAction(InputAction action) {
+    return toggleTypeByInputAction[action];
+}
+
+// TODO(captainurist): these should be localizable?
+MM_DEFINE_ENUM_SERIALIZATION_FUNCTIONS(PlatformKey, CASE_INSENSITIVE, {
+    {PlatformKey::KEY_F1,           "F1" },
+    {PlatformKey::KEY_F2,           "F2" },
+    {PlatformKey::KEY_F3,           "F3" },
+    {PlatformKey::KEY_F4,           "F4" },
+    {PlatformKey::KEY_F5,           "F5" },
+    {PlatformKey::KEY_F6,           "F6" },
+    {PlatformKey::KEY_F7,           "F7" },
+    {PlatformKey::KEY_F8,           "F8" },
+    {PlatformKey::KEY_F9,           "F9" },
+    {PlatformKey::KEY_F10,          "F10" },
+    {PlatformKey::KEY_F11,          "F11" },
+    {PlatformKey::KEY_F12,          "F12" },
+    {PlatformKey::KEY_F13,          "F13" },
+    {PlatformKey::KEY_F14,          "F14" },
+    {PlatformKey::KEY_F15,          "F15" },
+    {PlatformKey::KEY_F16,          "F16" },
+    {PlatformKey::KEY_F17,          "F17" },
+    {PlatformKey::KEY_F18,          "F18" },
+    {PlatformKey::KEY_F19,          "F19" },
+    {PlatformKey::KEY_F20,          "F20" },
+    {PlatformKey::KEY_F21,          "F21" },
+    {PlatformKey::KEY_F22,          "F22" },
+    {PlatformKey::KEY_F23,          "F23" },
+    {PlatformKey::KEY_F24,          "F24" },
+
+    {PlatformKey::KEY_A,            "A" },
+    {PlatformKey::KEY_B,            "B" },
+    {PlatformKey::KEY_C,            "C" },
+    {PlatformKey::KEY_D,            "D" },
+    {PlatformKey::KEY_E,            "E" },
+    {PlatformKey::KEY_F,            "F" },
+    {PlatformKey::KEY_G,            "G" },
+    {PlatformKey::KEY_H,            "H" },
+    {PlatformKey::KEY_I,            "I" },
+    {PlatformKey::KEY_J,            "J" },
+    {PlatformKey::KEY_K,            "K" },
+    {PlatformKey::KEY_L,            "L" },
+    {PlatformKey::KEY_M,            "M" },
+    {PlatformKey::KEY_N,            "N" },
+    {PlatformKey::KEY_O,            "O" },
+    {PlatformKey::KEY_P,            "P" },
+    {PlatformKey::KEY_Q,            "Q" },
+    {PlatformKey::KEY_R,            "R" },
+    {PlatformKey::KEY_S,            "S" },
+    {PlatformKey::KEY_T,            "T" },
+    {PlatformKey::KEY_U,            "U" },
+    {PlatformKey::KEY_V,            "V" },
+    {PlatformKey::KEY_W,            "W" },
+    {PlatformKey::KEY_X,            "X" },
+    {PlatformKey::KEY_Y,            "Y" },
+    {PlatformKey::KEY_Z,            "Z" },
+
+    {PlatformKey::KEY_DIGIT_1,       "1" },
+    {PlatformKey::KEY_DIGIT_2,       "2" },
+    {PlatformKey::KEY_DIGIT_3,       "3" },
+    {PlatformKey::KEY_DIGIT_4,       "4" },
+    {PlatformKey::KEY_DIGIT_5,       "5" },
+    {PlatformKey::KEY_DIGIT_6,       "6" },
+    {PlatformKey::KEY_DIGIT_7,       "7" },
+    {PlatformKey::KEY_DIGIT_8,       "8" },
+    {PlatformKey::KEY_DIGIT_9,       "9" },
+    {PlatformKey::KEY_DIGIT_0,       "0" },
+
+    {PlatformKey::KEY_UP,           "UP" },
+    {PlatformKey::KEY_DOWN,         "DOWN" },
+    {PlatformKey::KEY_LEFT,         "LEFT" },
+    {PlatformKey::KEY_RIGHT,        "RIGHT" },
+    {PlatformKey::KEY_RETURN,       "RETURN" },
+    {PlatformKey::KEY_ESCAPE,       "ESCAPE" },
+    {PlatformKey::KEY_SPACE,        "SPACE" },
+    {PlatformKey::KEY_PAGEDOWN,     "PAGE DOWN" },
+    {PlatformKey::KEY_PAGEUP,       "PAGE UP" },
+    {PlatformKey::KEY_TAB,          "TAB" },
+    {PlatformKey::KEY_SUBTRACT,     "SUBTRACT" },
+    {PlatformKey::KEY_ADD,          "ADD" },
+    {PlatformKey::KEY_END,          "END" },
+    {PlatformKey::KEY_DELETE,       "DELETE" },
+    {PlatformKey::KEY_HOME,         "HOME" },
+    {PlatformKey::KEY_INSERT,       "INSERT" },
+    {PlatformKey::KEY_COMMA,        "COMMA" },
+    {PlatformKey::KEY_DECIMAL,      "DECIMAL" },
+    {PlatformKey::KEY_SEMICOLON,    "SEMICOLON" },
+    {PlatformKey::KEY_PERIOD,       "PERIOD" },
+    {PlatformKey::KEY_SLASH,        "SLASH" },
+    {PlatformKey::KEY_SINGLEQUOTE,  "SQUOTE" },
+    {PlatformKey::KEY_BACKSLASH,    "BACKSLASH" },
+    {PlatformKey::KEY_BACKSPACE,    "BACKSPACE" },
+    {PlatformKey::KEY_LEFTBRACKET,  "L BRACKET" },
+    {PlatformKey::KEY_RIGHTBRACKET, "R BRACKET" },
+    {PlatformKey::KEY_TILDE,        "~" },
+    {PlatformKey::KEY_PRINTSCREEN,  "PRINT SCREEN" },
+    {PlatformKey::KEY_SELECT,       "SELECT" },
+
+    {PlatformKey::KEY_CONTROL,      "CONTROL"},
+    {PlatformKey::KEY_ALT,          "ALT"},
+    {PlatformKey::KEY_SHIFT,        "SHIFT"},
+
+    {PlatformKey::KEY_NUMPAD_0,      "NUMPAD 0" },
+    {PlatformKey::KEY_NUMPAD_1,      "NUMPAD 1" },
+    {PlatformKey::KEY_NUMPAD_2,      "NUMPAD 2" },
+    {PlatformKey::KEY_NUMPAD_3,      "NUMPAD 3" },
+    {PlatformKey::KEY_NUMPAD_4,      "NUMPAD 4" },
+    {PlatformKey::KEY_NUMPAD_5,      "NUMPAD 5" },
+    {PlatformKey::KEY_NUMPAD_6,      "NUMPAD 6" },
+    {PlatformKey::KEY_NUMPAD_7,      "NUMPAD 7" },
+    {PlatformKey::KEY_NUMPAD_8,      "NUMPAD 8" },
+    {PlatformKey::KEY_NUMPAD_9,      "NUMPAD 9" },
+
+    {PlatformKey::KEY_GAMEPAD_A,        "A (CROSS)"},
+    {PlatformKey::KEY_GAMEPAD_B,        "B (CIRCLE)"},
+    {PlatformKey::KEY_GAMEPAD_X,        "X (SQUARE)"},
+    {PlatformKey::KEY_GAMEPAD_Y,        "Y (TRIANGLE)"},
+    {PlatformKey::KEY_GAMEPAD_LEFT,     "DPAD LEFT"},
+    {PlatformKey::KEY_GAMEPAD_RIGHT,    "DPAD RIGHT"},
+    {PlatformKey::KEY_GAMEPAD_UP,       "DPAD UP"},
+    {PlatformKey::KEY_GAMEPAD_DOWN,     "DPAD DOWN"},
+    {PlatformKey::KEY_GAMEPAD_L1,       "L1"},
+    {PlatformKey::KEY_GAMEPAD_R1,       "R1"},
+    {PlatformKey::KEY_GAMEPAD_L3,       "L3"},
+    {PlatformKey::KEY_GAMEPAD_R3,       "R3"},
+    {PlatformKey::KEY_GAMEPAD_START,    "START"},
+    {PlatformKey::KEY_GAMEPAD_BACK,     "BACK"},
+    {PlatformKey::KEY_GAMEPAD_GUIDE,    "GUIDE"},
+    {PlatformKey::KEY_GAMEPAD_TOUCHPAD, "TOUCHPAD"},
+
+    {PlatformKey::KEY_GAMEPAD_LEFTSTICK_LEFT,   "LSTICK LEFT"},
+    {PlatformKey::KEY_GAMEPAD_LEFTSTICK_RIGHT,  "LSTICK RIGHT"},
+    {PlatformKey::KEY_GAMEPAD_LEFTSTICK_UP,     "LSTICK UP"},
+    {PlatformKey::KEY_GAMEPAD_LEFTSTICK_DOWN,   "LSTICK DOWN"},
+    {PlatformKey::KEY_GAMEPAD_RIGHTSTICK_LEFT,  "RSTICK LEFT"},
+    {PlatformKey::KEY_GAMEPAD_RIGHTSTICK_RIGHT, "RSTICK RIGHT"},
+    {PlatformKey::KEY_GAMEPAD_RIGHTSTICK_UP,    "RSTICK UP"},
+    {PlatformKey::KEY_GAMEPAD_RIGHTSTICK_DOWN,  "RSTICK DOWN"},
+    {PlatformKey::KEY_GAMEPAD_L2,               "L2"},
+    {PlatformKey::KEY_GAMEPAD_R2,               "R2"},
+    {PlatformKey::KEY_NONE,                     ""},
+    {PlatformKey::KEY_NONE,                     "-NOT-SET-"} // Allowed for deserialization only, relied on in TryParseDisplayName
+})
+
+std::string GetDisplayName(PlatformKey key) {
+    if (key == PlatformKey::KEY_NONE)
+        return "-NOT-SET-";
+    return toString(key); // Should always work.
+}
+
+bool TryParseDisplayName(std::string_view displayName, PlatformKey *outKey) {
+    return tryDeserialize(displayName, outKey);
+}
